// src/models/AiChatBriefEvent.ts
export type AiChatBriefEvent = { type: 'chat-client-request' } & AiChatRequestBriefEvent | { type: 'chat-client-response' } & AiChatResponseBriefEvent | { type: 'compactor-warning' } & AiChatCompactorWarningBriefEvent | { type: 'compressor-context-end' } & AiChatCompressorContextEndBriefEvent | { type: 'compressor-context-start' } & AiChatCompressorContextStartBriefEvent | { type: 'end' } & AiChatEndBriefEvent | { type: 'error' } & AiChatErrorBriefEvent | { type: 'manual-interrupt' } & AiChatManualInterruptBriefEvent | { type: 'message' } & AiChatMessageBriefEvent | { type: 'start' } & AiChatStartBriefEvent | { type: 'think' } & AiChatThinkBriefEvent | { type: 'tool-execution' } & AiChatToolExecutionBriefEvent | { type: 'tool-execution-awaiting-user-input' } & AiChatAwaitingInputBriefEvent | { type: 'tool-execution-sub-agent-call' } & AiChatSubAgentCallBriefEvent | { type: 'user-query' } & AiChatUserQueryBriefEvent;