// src/models/ToolExtension.ts
export type ToolExtension = { category: 'askUserQuestion' } & AskUserQuestionExtensionToolExtension | { category: 'imageGeneration' } & ImageGenerationExtensionToolExtension | { category: 'jsRunScript' } & JsRunScriptExtensionToolExtension | { category: 'jsRunScriptResult' } & JsRunScriptResultExtensionToolExtension | { category: 'mathFormula' } & MathFormulaExtensionToolExtension | { category: 'mathResult' } & MathResultExtensionToolExtension | { category: 'planApproval' } & PlanApprovalExtensionToolExtension | { category: 'skillResource' } & SkillResourceExtensionToolExtension | { category: 'sqlChartResult' } & SqlChartResultExtensionToolExtension | { category: 'sqlQuery' } & SqlQueryExtensionToolExtension | { category: 'sqlQueryResult' } & SqlQueryResultExtensionToolExtension | { category: 'taskProgress' } & TaskProgressExtensionToolExtension;